对 int、float、bool 类型配置参数值进行类型转换 (#448)
* redis->connect添加对port参数int转换

* 回滚

* 修正代码，timeout 也转换类型

* 对更多 int、float、bool 类型配置参数值进行类型转换

* 修复测试

+++ src/Components/swoole/src/Server/TcpServer/Server.php
@@ -55,7 +55,7 @@ class Server extends Base implements ISwooleTcpServer
     protected function createServer(): void
     {
         $config = $this->getServerInitConfig();
-        $this->swooleServer = new \Swoole\Server($config['host'], $config['port'], $config['mode'], $config['sockType']);
+        $this->swooleServer = new \Swoole\Server($config['host'], (int) $config['port'], (int) $config['mode'], (int) $config['sockType']);
     }
 
     /**
@@ -78,9 +78,9 @@ class Server extends Base implements ISwooleTcpServer
     {
         return [
             'host'      => $this->config['host'] ?? '0.0.0.0',
-            'port'      => $this->config['port'] ?? 8080,
+            'port'      => (int) ($this->config['port'] ?? 8080),
             'sockType'  => isset($this->config['sockType']) ? (\SWOOLE_SOCK_TCP | $this->config['sockType']) : \SWOOLE_SOCK_TCP,
-            'mode'      => $this->config['mode'] ?? \SWOOLE_BASE,
+            'mode'      => (int) ($this->config['mode'] ?? \SWOOLE_BASE),
         ];
     }
